+++ fugue/src/main/java/com/atlassian/fugue/Functions.java
@@ -743,4 +743,15 @@ public class Functions {
       }
     }
   }
+
+  static <A> Predicate<A> countingPredicate(final int n){
+    if(n < 0){
+      throw new IllegalArgumentException("n must be positive");
+    }
+    final int[] seen = {0};
+    return a -> {
+      seen[0] = seen[0] + 1;
+      return seen[0] <= n;
+    };
+  }
 }